fontを最新のものに統一、クリア時のスコア計算方法を変更

+++ MODEL77Framework/Assets/G20/Scripts/PlayerInput/G20_BulletShooter.cs
@@ -86,7 +86,7 @@ public class G20_BulletShooter : G20_Singleton<G20_BulletShooter>
             {
                 AssistHitCount++;
             }
-            HitRate = AssistHitCount/ShotCount;
+            HitRate = (float)AssistHitCount/(float)ShotCount;
 
             ShootPanel((Vector2)shootPoint);
